Delete Image From the upload folder

+++ app/Http/Controllers/Admin/SliderController.php
@@ -111,8 +111,10 @@ class SliderController extends Controller
     }
     public function destroy($id){
         $slider= Slider::find($id);
-        unlink('uploads/slider/'.$slider->image);
 
+        if(file_exists('uploads/slider/'.$slider->image)) {
+            unlink('uploads/slider/' . $slider->image);
+        }
         $slider->delete();
 
         return redirect()->back()->with([
